# Fix #5 by using as many BigInt in-place ops as possible to avoid allo…