--- a/benchmarks/floats.jl
+++ b/benchmarks/floats.jl
@@ -51,9 +51,10 @@ end
 
 
 function prof(n)
-    io = IOBuffer("3.2925339999999996e-18")
+    io = IOBuffer("2.2250738585072011e-308")
+    res = Parsers.Result(Float64)
     for i = 1:n
         seekstart(io)
-        Main.Parsers.xparse(io, Float64)
+        r = Parsers.defaultparser(io, res)
     end
 end