--- a/test/floats.jl
+++ b/test/floats.jl
@@ -338,8 +338,6 @@ r = Parsers.defaultparser(IOBuffer("1.000000000000000111022302462515654042363166
 r = Parsers.defaultparser(IOBuffer("1.00000000000000011102230246251565404236316680908203126"), Parsers.Result(Float64))
 @test r.result === 1.0000000000000002
 @test r.code === OK | EOF
-# # Slightly higher, but you have to read all the way to the end.
-# r = Parsers.defaultparser(IOBuffer("1.00000000000000011102230246251565404236316680908203125"), Parsers.Result(Float64))
 
 r = Parsers.defaultparser(IOBuffer("-5.871153289887625082e-01"), Parsers.Result(Float64))
 @test r.result === -5.871153289887625082e-01
@@ -373,6 +371,363 @@ r = Parsers.defaultparser(IOBuffer(".24409E+03"), Parsers.Result(Float64))
 
 r = Parsers.defaultparser(IOBuffer(".24409E+03 "), Parsers.Result(Float64))
 @test r.result === 244.09
+@test r.code === OK | EOF
+
+# from https://www.icir.org/vern/papers/testbase-report.pdf
+float = 5.0 * exp10(+125)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 69.0 * exp10(+267)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 999.0 * exp10(-26)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 7861.0 * exp10(-34)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 75569.0 * exp10(-254)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 928609.0 * exp10(-261)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 9210917.0 * exp10(+80)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 84863171.0 * exp10(+114)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 653777767.0 * exp10(+273)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 5232604057.0 * exp10(-298)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 27235667517.0 * exp10(-109)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 653532977297.0 * exp10(-123)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 3142213164987.0 * exp10(-294)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 46202199371337.0 * exp10(-72)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 231010996856685.0 * exp10(-73)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 9324754620109615.0 * exp10(+212)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 78459735791271921.0 * exp10(+49)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 272104041512242479.0 * exp10(+200)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 6802601037806061975.0 * exp10(+198)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 20505426358836677347.0 * exp10(-221)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 836168422905420598437.0 * exp10(-234)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 4891559871276714924261.0 * exp10(+222)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+
+float = 9.0 * exp10(-265)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 85.0 * exp10(-037)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 623.0 * exp10(+100)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 3571.0 * exp10(+263)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 81661.0 * exp10(+153)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 920657.0 * exp10(-023)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 4603285.0 * exp10(-024)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 87575437.0 * exp10(-309)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 245540327.0 * exp10(+122)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 6138508175.0 * exp10(+120)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 83356057653.0 * exp10(+193)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 619534293513.0 * exp10(+124)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 2335141086879.0 * exp10(+218)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 36167929443327.0 * exp10(-159)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 609610927149051.0 * exp10(-255)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 3743626360493413.0 * exp10(-165)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 94080055902682397.0 * exp10(-242)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 899810892172646163.0 * exp10(+283)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 7120190517612959703.0 * exp10(+120)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 25188282901709339043.0 * exp10(-252)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 308984926168550152811.0 * exp10(-052)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 6372891218502368041059.0 * exp10(+064)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+
+float = 8511030020275656.0 * exp2(-0342)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 5201988407066741.0 * exp2(-0824)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 6406892948269899.0 * exp2(+0237)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 8431154198732492.0 * exp2(+0072)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 6475049196144587.0 * exp2(+0099)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 8274307542972842.0 * exp2(+0726)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 5381065484265332.0 * exp2(-0456)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 6761728585499734.0 * exp2(-1057)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 7976538478610756.0 * exp2(+0376)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 5982403858958067.0 * exp2(+0377)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 5536995190630837.0 * exp2(+0093)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 7225450889282194.0 * exp2(+0710)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 7225450889282194.0 * exp2(+0709)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 8703372741147379.0 * exp2(+0117)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 8944262675275217.0 * exp2(-1001)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 7459803696087692.0 * exp2(-0707)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 6080469016670379.0 * exp2(-0381)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 8385515147034757.0 * exp2(+0721)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 7514216811389786.0 * exp2(-0828)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 8397297803260511.0 * exp2(-0345)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 6733459239310543.0 * exp2(+0202)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 8091450587292794.0 * exp2(-0473)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+
+float = 6567258882077402.0 * exp2(+952)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 6712731423444934.0 * exp2(+535)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 6712731423444934.0 * exp2(+534)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 5298405411573037.0 * exp2(-957)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 5137311167659507.0 * exp2(-144)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 6722280709661868.0 * exp2(+363)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 5344436398034927.0 * exp2(-169)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 8369123604277281.0 * exp2(-853)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 8995822108487663.0 * exp2(-780)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 8942832835564782.0 * exp2(-383)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 8942832835564782.0 * exp2(-384)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 8942832835564782.0 * exp2(-385)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 6965949469487146.0 * exp2(-249)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 6965949469487146.0 * exp2(-250)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 6965949469487146.0 * exp2(-251)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 7487252720986826.0 * exp2(+548)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 5592117679628511.0 * exp2(+164)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 8887055249355788.0 * exp2(+665)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 6994187472632449.0 * exp2(+690)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 8797576579012143.0 * exp2(+588)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 7363326733505337.0 * exp2(+272)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
+@test r.code === OK | EOF
+float = 8549497411294502.0 * exp2(-448)
+r = Parsers.defaultparser(IOBuffer(string(float)), Parsers.Result(Float64))
+@test r.result === float
 @test r.code === OK | EOF
 
 end # @testset