@quinnj 2019-03-07
@simonbyrne, just pinging you here if you wouldn't mind taking a look; I know we've talked strategy here a bit and I think this is looking pretty good. Here's a taste of the perf improvements by pre-allocating a few BigInts and doing in-place operations:

Before:
```julia
julia> run(@benchmarkable Parsers.defaultparser(io, r) setup=(io = IOBuffer("0.0017138347201173243"); r = Parsers.Result($T)))
BenchmarkTools.Trial:
  memory estimate:  408 bytes
  allocs estimate:  24
  --------------
  minimum time:     1.098 μs (0.00% GC)
  median time:      1.277 μs (0.00% GC)
  mean time:        1.363 μs (0.00% GC)
  maximum time:     29.360 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

After:
```julia
julia> run(@benchmarkable Parsers.defaultparser(io, r) setup=(io = IOBuffer("0.0017138347201173243"); r = Parsers.Result($T)))
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     216.000 ns (0.00% GC)
  median time:      232.000 ns (0.00% GC)
  mean time:        244.993 ns (0.00% GC)
  maximum time:     36.176 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

And compared to Base.parse:
```julia
julia> @benchmark parse(Float64, "0.0017138347201173243")
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     299.788 ns (0.00% GC)
  median time:      310.359 ns (0.00% GC)
  mean time:        329.445 ns (0.00% GC)
  maximum time:     928.122 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     245
```