@simonbyrne 2019-03-07
Nice. It generally looks okay. The main problem is handling multithreading.

For worst-case test values, see https://www.icir.org/vern/papers/testbase-report.pdf