@quinnj 2019-03-08
@simonbyrne thanks for taking a look; good call on the multithreaded case, we should be good now. I also added a bunch of tests from the paper you linked; great stuff!